Docker/sinica_artplace/requirements.txt
asgiref==3.3.4
Django==3.2.3
django-rest-framework==0.1.0
djangorestframework==3.12.4
django-ckeditor
gunicorn==20.1.0
Pillow==8.2.0
pytz==2021.1
sqlparse==0.4.1
celery
redis
django-celery-results
django-celery-beat